test: rename and update transition tests for full-site ClientRouter
